% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outLyapFD-data.R
\docType{data}
\name{outLyapFD}
\alias{outLyapFD}
\title{An output of FDLyapu run}
\format{An object of class \code{list} of length 2.}
\usage{
outLyapFD
}
\description{
This data set provides two examples of outputs (Rossler-1976 and Rossler-1979 systems) produced by the FDLyapu algorithm.
}
\author{
Sylvain Mangiarotti \email{sylvain.mangiarotti@cesbio.cnes.fr}
}
\keyword{data}
