% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reit98-data.R
\docType{data}
\name{Reit98}
\alias{Reit98}
\title{The 9D model for a Rayleigh-Bénard convection in a square cell}
\format{An object of class \code{matrix} with 55 rows and 9 columns.}
\usage{
Reit98
}
\description{
The 9D model for a Rayleigh-Bénard convection in a square cell
was introduced by Reiterer et al. in 1998. It can produce an
apparently smooth transition from chaos to hyperchaos.
}
\references{
P. Reiterer, C. Lainscsek, F. Schürrer, C. Letellier, and J. Maquet,
A nine-dimensional Lorenz system to study high-dimensional chaos,
Journal of Physics A, 31, 7121-7139, 1998.
}
\author{
Sylvain Mangiarotti \email{sylvain.mangiarotti@cesbio.cnes.fr}
}
\keyword{data}
